refactoring of control

See merge request compsci308_2020spring/parser_parser_team17!117

+++ src/backEnd/commands/IfElse.java
@@ -6,14 +6,14 @@ import java.util.LinkedList;
 public class IfElse extends Command{
 
   private final int number =1;
-  private int check;
+  private double check;
   public IfElse(){
     super();
     super.numberOfArgs=number;
   }
   public IfElse(LinkedList<String> varags, Control control){
     super(varags,control);
-    check = Integer.parseInt(varags.get(0));
+    check = Double.parseDouble(varags.get(0));
   }
 
   @Override
